Fix mypy issues

+++ game/settings/settings.py
@@ -545,7 +545,7 @@ class Settings:
         max=150,
     )
 
-    unit_multiplier: int = bounded_float_option(
+    unit_multiplier: float = bounded_float_option(
         "Desired unit multiplier",
         page=MISSION_GENERATOR_PAGE,
         section=GAMEPLAY_SECTION,
